requirements.txt
pydantic-settings
redis
urllib3
requests
celery
python-libnmap